Update poetry lock file format.
I wanted to add a package, but that rewrote the whole lock file to a new
format as a side-effect. That's fine, but it makes for a messy diff.
This commit is the result of performing a dummy update:

    poetry run poetry update non_existent_package

just to update the lock file to the new format, with no actual changes in
the dependencies.

That was the intention, anyway. Looking at the diff, some versions
were bumped, after all. I don't understand why, but it seems harmless.